requirements.txt
streamlit
google-generativeai
python-dotenv
reportlab
pandas
sentence-transformers
chromadb
pysqlite3-binary
